Type conversions to stop errors when switching data formats

+++ External/gem/GemSlider.m
@@ -68,7 +68,7 @@ classdef GemSlider < GemUserInterfaceObject
         end
         
         function SetSliderSteps(obj, steps)
-            obj.SliderSteps = steps;
+            obj.SliderSteps = double(steps);
             if obj.ComponentHasBeenCreated
                 set(obj.GraphicalComponentHandle, 'SliderStep', steps);
             end
